pyproject.toml
[tool.poetry]
name = "camel-oasis"
version = "0.2.3"
description = "Open Agents Social Interaction Simulations on a Large Scale"
authors = ["CAMEL-AI.org"]
readme = "README.md"
license = "Apache License 2.0"
keywords = [
    "recommendation-system",
    "communicative-ai",
    "ai-societies",
    "social-simulations",
    "artificial-intelligence",
    "deep-learning",
    "multi-agent-systems",
    "natural-language-processing",
    "large-language-models",
]
packages = [
    { include = "oasis" },
]
repository = "https://github.com/camel-ai/oasis.git"

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
pandas = "2.2.2"
igraph = "0.11.6"
cairocffi = "1.7.1"
pytest = "8.2.0"
pytest-asyncio = "0.23.6"
pillow = "10.3.0"
unstructured = "0.13.7"
sentence-transformers = "3.0.0"
pre-commit = "3.7.1"
prance = "23.6.21.0"
openapi-spec-validator = "0.7.1"
slack_sdk = "3.31.0"
neo4j = "5.23.0"
camel-ai = "0.2.70"
requests_oauthlib = "2.0.0"
requests = "^2.32.3"
PyYAML = "^6.0.2"
python-dotenv = "^1.0.0"
google-genai = "^0.3.0"
tenacity = "^9.0.0"
aiohttp = "^3.13.2"
pyarrow = "^22.0.0"
tqdm = "^4.67.1"

[tool.poetry.group.dev.dependencies]
matplotlib = "^3.10.7"

[tool.poetry.group.open_models_testing]
optional = true

[tool.poetry.group.open_models_testing.dependencies]
transformers = "^4.45.0"
torch = "^2.4.1"
pydantic = "^2.9.2"
tqdm = "^4.66.5"
llama-cpp-python = "^0.3.2"
openai = "^1.51.2"
huggingface_hub = "^0.34.0"
requests = "^2.32.3"
accelerate = "^1.1.1"
protobuf = "^5.29.1"
sentencepiece = "^0.2.0"
torchvision = "^0.19.1"
diffusers = { git = "https://github.com/huggingface/diffusers.git", rev = "main" }
safetensors = "^0.4.5"

[[tool.poetry.source]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "supplemental"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
